deploy-agents/requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
playwright>=1.40.0
